Use simpler rules in cascadingfilter

+++ processor/cascadingfilterprocessor/sampling/policy.go
@@ -35,7 +35,7 @@ type TraceData struct {
 	// Decisiontime time when sampling decision was taken.
 	DecisionTime time.Time
 	// SpanCount track the number of spans on the trace.
-	SpanCount int64
+	SpanCount int32
 	// ReceivedBatches stores all the batches received for the trace.
 	ReceivedBatches []pdata.Traces
 }
